chore: Move status, scraperEnum to store directory & delete groupContext
