+++ packages/admin/src/store/scraperEnum.ts
@@ -1,0 +1,6 @@
+export enum Scrapers {
+  Notice = "학과별 공지사항",
+  StudentRestaurant = "힉생 식당",
+  DomitoryRestaurant = "기숙사 식당",
+  CollegeSchedule = "학사일정",
+}
