+++ packages/admin/src/store/statusEnum.ts
@@ -1,0 +1,6 @@
+export enum Status {
+  All = "모두",
+  Running = "실행중",
+  Waiting = "대기중",
+  Error = "장애",
+}
